requirements.txt
functions-framework==3.*
requests
schedule
google-cloud-storage
gunicorn
google-cloud-logging
pandas
python-dotenv
setuptools